fix: error codes need to be uint32_t
Changing the 'ffa_error' function to cast to uint32_t the
argument where the error code is placed.
This is because the FF-A specification mandates the 32-bit
value in the w2/x2 register of the ABI.
The uint32_t explicit cast guarantees the value is zero
extended.

+++ inc/hf/ffa_internal.h
@@ -17,9 +17,15 @@
 
 #define FFA_VERSION_RESERVED_BIT (UINT32_C(1) << 31)
 
+/**
+ * The helper functions prepares the FFA_ERROR 32-bit ABI.
+ * The enum is cast to a uint32_t because it must be zero-extended to 64 bits,
+ * not sign-extended.
+ */
 static inline struct ffa_value ffa_error(enum ffa_error error_code)
 {
-	return (struct ffa_value){.func = FFA_ERROR_32, .arg2 = error_code};
+	return (struct ffa_value){.func = FFA_ERROR_32,
+				  .arg2 = (uint32_t)error_code};
 }
 
 struct ffa_value ffa_msg_recv_return(const struct vm *receiver);
